refactor(proxy): extract api-gate module from proxy.ts
proxy.ts's /api/ branch (multipart allowlist, public-route
classification, Origin/CSRF check, HMAC verification) was interleaved
with locale routing, CSP, and SEO header logic in one 916-line
function, even though three existing test files already exercised it
as a separate concern (proxy-route-guards, proxy-public-api-patterns,
proxy-origin-allowed).

Extracted the whole /api/ branch into utils/api-gate.ts as
gateApiRequest(request), along with isDev, the origin-allowlist
helpers, and the public-route constants it depends on (moving isDev
too avoids a circular import back into proxy.ts, since it's used both
inside and outside the extracted region). proxy.ts re-exports the same
six symbols so the three test files needed zero changes; verified all
120 proxy-suite tests and the full 2219-test suite pass unchanged.

proxy.ts drops from 916 to 573 lines. Locale routing, CSP, and SEO
headers are deliberately left untouched — no test evidence supports
extracting those as separate modules yet.
